optimize router, remove some not implement feature structs
