+++ pkg/api/v2/config.go
@@ -83,10 +83,8 @@ type DelayInjectConfig struct {
 type RouterConfig struct {
 	Match           RouterMatch            `json:"match"`
 	Route           RouteAction            `json:"route"`
-	Redirect        RedirectAction         `json:"redirect"`
 	DirectResponse  *DirectResponseAction  `json:"direct_response"`
 	MetadataConfig  MetadataConfig         `json:"metadata"`
-	Decorator       Decorator              `json:"decorator"`
 	PerFilterConfig map[string]interface{} `json:"per_filter_config"`
 }
