+++ pkg/api/v2/unmarshal_test.go
@@ -308,10 +308,7 @@ func TestRouterConfigUmaeshal(t *testing.T) {
 		vh := router.VirtualHosts[0]
 		if !(vh.Name != "virtual" &&
 			len(vh.Domains) == 1 &&
-			vh.Domains[0] == "*" &&
-			len(vh.VirtualClusters) == 1 &&
-			vh.VirtualClusters[0].Name == "vc" &&
-			vh.VirtualClusters[0].Pattern == "test") {
+			vh.Domains[0] == "*") {
 			t.Error("virtual host failed")
 		}
 		if len(vh.Routers) != 1 {
@@ -319,8 +316,6 @@ func TestRouterConfigUmaeshal(t *testing.T) {
 		} else {
 			router := vh.Routers[0]
 			if !(router.Match.Prefix == "/" &&
-				router.Match.Runtime.DefaultValue == 10 &&
-				router.Match.Runtime.RuntimeKey == "test" &&
 				len(router.Match.Headers) == 1 &&
 				router.Match.Headers[0].Name == "service" &&
 				router.Match.Headers[0].Value == "test") {
@@ -334,8 +329,7 @@ func TestRouterConfigUmaeshal(t *testing.T) {
 				router.Route.RetryPolicy.RetryTimeout == time.Minute &&
 				router.Route.RetryPolicy.RetryOn == true &&
 				router.Route.RetryPolicy.NumRetries == 10 &&
-				reflect.DeepEqual(meta, router.Metadata) &&
-				router.Decorator == "test") {
+				reflect.DeepEqual(meta, router.Metadata)) {
 				t.Error("virtual host failed")
 			}
 			if len(router.Route.WeightedClusters) != 1 {
@@ -347,10 +341,6 @@ func TestRouterConfigUmaeshal(t *testing.T) {
 					reflect.DeepEqual(meta, wc.Cluster.MetadataMatch)) {
 					t.Error("virtual host failed")
 				}
-			}
-			if !(router.Redirect.HostRedirect == "test" &&
-				router.Redirect.ResponseCode == 302) {
-				t.Error("virtual host failed")
 			}
 
 		}
