+++ pkg/router/base_rule_test.go
@@ -20,145 +20,12 @@ package router
 import (
 	"math/rand"
 	"reflect"
-	"regexp"
 	"testing"
 
 	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/protocol"
 	"github.com/alipay/sofa-mosn/pkg/types"
 )
-
-func TestPrefixRouteRuleImpl(t *testing.T) {
-	virtualHostImpl := &VirtualHostImpl{virtualHostName: "test"}
-	testCases := []struct {
-		prefix     string
-		headerpath string
-		expected   bool
-	}{
-		{"/", "/", true},
-		{"/", "/test", true},
-		{"/", "/test/foo", true},
-		{"/", "/foo?key=value", true},
-		{"/foo", "/foo", true},
-		{"/foo", "/footest", true},
-		{"/foo", "/foo/test", true},
-		{"/foo", "/foo?key=value", true},
-		{"/foo", "/", false},
-		{"/foo", "/test", false},
-	}
-	for i, tc := range testCases {
-		route := &v2.Router{
-			RouterConfig: v2.RouterConfig{
-				Match: v2.RouterMatch{Prefix: tc.prefix},
-				Route: v2.RouteAction{
-					RouterActionConfig: v2.RouterActionConfig{
-						ClusterName: "test",
-					},
-				},
-			},
-		}
-		routuRule, _ := NewRouteRuleImplBase(virtualHostImpl, route)
-		rr := &PrefixRouteRuleImpl{
-			routuRule,
-			route.Match.Prefix,
-		}
-		headers := protocol.CommonHeader(map[string]string{protocol.MosnHeaderPathKey: tc.headerpath})
-		result := rr.Match(headers, 1)
-		if (result != nil) != tc.expected {
-			t.Errorf("#%d want matched %v, but get matched %v\n", i, tc.expected, result)
-		}
-		if result != nil {
-			if result.RouteRule().PathMatchCriterion().MatchType() != types.Prefix {
-				t.Errorf("#%d match type is not expected", i)
-			}
-		}
-	}
-}
-
-func TestPathRouteRuleImpl(t *testing.T) {
-	virtualHostImpl := &VirtualHostImpl{virtualHostName: "test"}
-	testCases := []struct {
-		path          string
-		headerpath    string
-		caseSensitive bool //no interface to set caseSensitive, need hack
-		expected      bool
-	}{
-		{"/test", "/test", false, true},
-		{"/test", "/Test", false, true},
-		{"/test", "/Test", true, false},
-		{"/test", "/test/test", false, false},
-	}
-	for i, tc := range testCases {
-		route := &v2.Router{
-			RouterConfig: v2.RouterConfig{
-				Match: v2.RouterMatch{Path: tc.path},
-				Route: v2.RouteAction{
-					RouterActionConfig: v2.RouterActionConfig{
-						ClusterName: "test",
-					},
-				},
-			},
-		}
-		base, _ := NewRouteRuleImplBase(virtualHostImpl, route)
-		base.caseSensitive = tc.caseSensitive //hack case sensitive
-		rr := &PathRouteRuleImpl{base, route.Match.Path}
-		headers := protocol.CommonHeader(map[string]string{protocol.MosnHeaderPathKey: tc.headerpath})
-		result := rr.Match(headers, 1)
-		if (result != nil) != tc.expected {
-			t.Errorf("#%d want matched %v, but get matched %v\n", i, tc.expected, result)
-		}
-		if result != nil {
-			if result.RouteRule().PathMatchCriterion().MatchType() != types.Exact {
-				t.Errorf("#%d match type is not expected", i)
-			}
-		}
-
-	}
-}
-
-func TestRegexRouteRuleImpl(t *testing.T) {
-	virtualHostImpl := &VirtualHostImpl{virtualHostName: "test"}
-	testCases := []struct {
-		regexp     string
-		headerpath string
-		expected   bool
-	}{
-		{".*", "/", true},
-		{".*", "/path", true},
-		{"/[0-9]+", "/12345", true},
-		{"/[0-9]+", "/test", false},
-	}
-	for i, tc := range testCases {
-		route := &v2.Router{
-			RouterConfig: v2.RouterConfig{
-				Match: v2.RouterMatch{Regex: tc.regexp},
-				Route: v2.RouteAction{
-					RouterActionConfig: v2.RouterActionConfig{
-						ClusterName: "test",
-					},
-				},
-			},
-		}
-		re := regexp.MustCompile(tc.regexp)
-		routuRule, _ := NewRouteRuleImplBase(virtualHostImpl, route)
-
-		rr := &RegexRouteRuleImpl{
-			routuRule,
-			route.Match.Regex,
-			re,
-		}
-		headers := protocol.CommonHeader(map[string]string{protocol.MosnHeaderPathKey: tc.headerpath})
-		result := rr.Match(headers, 1)
-		if (result != nil) != tc.expected {
-			t.Errorf("#%d want matched %v, but get matched %v\n", i, tc.expected, result)
-		}
-		if result != nil {
-			if result.RouteRule().PathMatchCriterion().MatchType() != types.Regex {
-				t.Errorf("#%d match type is not expected", i)
-			}
-		}
-	}
-}
 
 func TestWeightedClusterSelect(t *testing.T) {
 	routerMock1 := &v2.Router{}
@@ -191,7 +58,6 @@ func TestWeightedClusterSelect(t *testing.T) {
 			},
 		},
 	}
-
 	routerMock2 := &v2.Router{}
 	routerMock2.Route = v2.RouteAction{
 		RouterActionConfig: v2.RouterActionConfig{
@@ -232,7 +98,6 @@ func TestWeightedClusterSelect(t *testing.T) {
 		routerCase: []*v2.Router{routerMock1, routerMock2},
 		ratio:      []uint{9, 1},
 	}
-
 	for index, routecase := range testCases.routerCase {
 		routeRuleImplBase, _ := NewRouteRuleImplBase(nil, routecase)
 		var dcCount, w1Count, w2Count uint
@@ -385,6 +250,7 @@ func Test_RouteRuleImplBase_FinalizeRequestHeaders(t *testing.T) {
 			},
 			want: protocol.CommonHeader{"host": "xxx.default.svc.cluster.local", "authority": "www.xxx.com", "level": "1,2,3", "route": "true", "vhost": "true", "global": "true"},
 		},
+
 		{
 			name: "case2",
 			args: args{
@@ -436,7 +302,6 @@ func Test_RouteRuleImplBase_FinalizeRequestHeaders(t *testing.T) {
 			want: protocol.CommonHeader{"host": "xxx.default.svc.cluster.local", "level": "1,3", "route": "true", "global": "true"},
 		},
 	}
-
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			tt.args.rri.FinalizeRequestHeaders(tt.args.headers, tt.args.requestInfo)
