+++ pkg/router/factory.go
@@ -19,6 +19,7 @@ package router
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/log"
@@ -34,11 +35,12 @@ var defaultRouterRuleFactoryOrder routerRuleFactoryOrder
 
 func RegisterRouterRule(f RouterRuleFactory, order uint32) {
 	if defaultRouterRuleFactoryOrder.order < order {
-		log.DefaultLogger.Infof("register router rule, order %d", order)
+		log.DefaultLogger.Infof(RouterLogFormat, "Extend", "RegisterRouterRule", fmt.Sprintf("order is %d", order))
 		defaultRouterRuleFactoryOrder.factory = f
 		defaultRouterRuleFactoryOrder.order = order
 	} else {
-		log.DefaultLogger.Warnf("current register order is %d, order %d register failed", defaultRouterRuleFactoryOrder.order, order)
+		msg := fmt.Sprintf("current register order is %d, order %d register failed", defaultRouterRuleFactoryOrder.order, order)
+		log.DefaultLogger.Errorf(RouterLogFormat, "Extend", "RegisterRouterRule", msg)
 	}
 }
 
@@ -58,11 +60,12 @@ var makeHandlerChainOrder handlerChainOrder
 
 func RegisterMakeHandlerChain(f MakeHandlerChain, order uint32) {
 	if makeHandlerChainOrder.order < order {
-		log.DefaultLogger.Infof("register make handler chain, order %d", order)
+		log.DefaultLogger.Infof(RouterLogFormat, "Extend", "RegisterHandlerChain", fmt.Sprintf("order is %d", order))
 		makeHandlerChainOrder.makeHandlerChain = f
 		makeHandlerChainOrder.order = order
 	} else {
-		log.DefaultLogger.Warnf("current register order is %d, order %d register failed", makeHandlerChainOrder.order, order)
+		msg := fmt.Sprintf("current register order is %d, order %d register failed", makeHandlerChainOrder.order, order)
+		log.DefaultLogger.Errorf(RouterLogFormat, "Extend", "RegisterHandlerChain", msg)
 	}
 }
